Implement global exception handler
Custom exceptions are also added:
* UserNotFoundException - 404
* BadTokenException - 400
* InternalServerException - 500 (should never arise)

+++ client/src/main/java/com/example/client/service/LecturerService.java
@@ -14,7 +14,7 @@ public interface LecturerService {
 
     Lecturer findByEmail(String email);
 
-    Lecturer findById(int id);
+    Lecturer findById(Long id);
 
     void changePassword(Lecturer lecturer, String newPassword);
